重新格式化代码

+++ Output/AEAssist/AEAssistLoader.cs
@@ -1,20 +1,14 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Drawing;
 using System.IO;
 using System.Reflection;
-using System.Threading.Tasks;
 using System.Windows.Media;
-using AEAssist;
 using Clio.Utilities;
 using ff14bot;
 using ff14bot.AClasses;
 using ff14bot.Enums;
 using ff14bot.Helpers;
-using ff14bot.Managers;
-using ff14bot.Navigation;
 using TreeSharp;
-using Core = ff14bot.Core;
 
 namespace AEAssist
 {
@@ -23,27 +17,12 @@ namespace AEAssist
         private const string ProjectName = "AEAssist";
         private const string ProjectAssemblyName = "AEAssist.dll";
 
-        public override string Name
-        {
-            get => ProjectName;
-        }
-
-        public override float PullRange
-        {
-            get => 25;
-        }
-
-        public override bool WantButton
-        {
-            get => true;
-        }
-
         private static readonly string ProjectAssembly = Path.Combine(Environment.CurrentDirectory,
             $@"Routines\{ProjectName}\{ProjectAssemblyName}");
 
         private static readonly string GreyMagicAssembly = Path.Combine(Environment.CurrentDirectory, @"GreyMagic.dll");
 
-        public static readonly HashSet<string> ExternelDlls = new HashSet<string>()
+        public static readonly HashSet<string> ExternelDlls = new HashSet<string>
         {
             "MaterialDesignColors",
             "MaterialDesignThemes.Wpf",
@@ -51,32 +30,111 @@ namespace AEAssist
             "MongoDB.Bson"
         };
 
+        public Dictionary<string, PropertyInfo> Behaviors = new Dictionary<string, PropertyInfo>();
+
+        public object Entry;
+
+        private bool Loaded;
+        public Dictionary<string, MethodInfo> Methods = new Dictionary<string, MethodInfo>();
+
+        public override string Name => ProjectName;
+
+        public override float PullRange => 25;
+
+        public override bool WantButton => true;
+
         public override ClassJobType[] Class
         {
             get
             {
-                switch (ff14bot.Core.Me.CurrentJob)
+                switch (Core.Me.CurrentJob)
                 {
                     case ClassJobType.Bard:
                     case ClassJobType.Reaper:
-                        return new[] {ff14bot.Core.Me.CurrentJob};
+                        return new[] {Core.Me.CurrentJob};
                     default:
                     {
-                        Logging.Write(Colors.Red, $@"[AEAssist] {ff14bot.Core.Me.CurrentJob} is not supported.");
-                        return new[] {ff14bot.Core.Me.CurrentJob};
+                        Logging.Write(Colors.Red, $@"[AEAssist] {Core.Me.CurrentJob} is not supported.");
+                        return new[] {Core.Me.CurrentJob};
                     }
                 }
             }
         }
 
-        public Dictionary<string, PropertyInfo> Behaviors = new Dictionary<string, PropertyInfo>();
-        public Dictionary<string, MethodInfo> Methods = new Dictionary<string, MethodInfo>();
+        private static string CompiledAssembliesPath => Path.Combine(Utilities.AssemblyDirectory, "CompiledAssemblies");
 
-        public object Entry;
+        public override Composite RestBehavior
+        {
+            get
+            {
+                if (!Behaviors.TryGetValue("RestBehavior", out var prop)) return default;
 
-        private bool Loaded;
+                return prop.GetValue(Entry, null) as Composite;
+            }
+        }
 
-        void LoadAsm()
+
+        public override Composite PreCombatBuffBehavior
+        {
+            get
+            {
+                if (!Behaviors.TryGetValue("PreCombatBuffBehavior", out var prop)) return default;
+
+                return prop.GetValue(Entry, null) as Composite;
+            }
+        }
+
+        public override Composite PullBehavior
+        {
+            get
+            {
+                if (!Behaviors.TryGetValue("PullBehavior", out var prop)) return default;
+
+                return prop.GetValue(Entry, null) as Composite;
+            }
+        }
+
+        public override Composite HealBehavior
+        {
+            get
+            {
+                if (!Behaviors.TryGetValue("HealBehavior", out var prop)) return default;
+
+                return prop.GetValue(Entry, null) as Composite;
+            }
+        }
+
+        public override Composite CombatBuffBehavior
+        {
+            get
+            {
+                if (!Behaviors.TryGetValue("CombatBuffBehavior", out var prop)) return default;
+
+                return prop.GetValue(Entry, null) as Composite;
+            }
+        }
+
+        public override Composite CombatBehavior
+        {
+            get
+            {
+                if (!Behaviors.TryGetValue("CombatBehavior", out var prop)) return default;
+
+                return prop.GetValue(Entry, null) as Composite;
+            }
+        }
+
+        public override Composite PullBuffBehavior
+        {
+            get
+            {
+                if (!Behaviors.TryGetValue("PullBuffBehavior", out var prop)) return default;
+
+                return prop.GetValue(Entry, null) as Composite;
+            }
+        }
+
+        private void LoadAsm()
         {
             RedirectAssembly();
             var path = ProjectAssembly;
@@ -106,7 +164,7 @@ namespace AEAssist
         {
             ResolveEventHandler handler = (sender, args) =>
             {
-                string name = Assembly.GetEntryAssembly().GetName().Name;
+                var name = Assembly.GetEntryAssembly().GetName().Name;
                 var requestedAssembly = new AssemblyName(args.Name);
                 return requestedAssembly.Name != name ? null : Assembly.GetEntryAssembly();
             };
@@ -132,24 +190,16 @@ namespace AEAssist
             AppDomain.CurrentDomain.AssemblyResolve += MaterialDesignHandler;
         }
 
-        static string GetAssPath(string name)
+        private static string GetAssPath(string name)
         {
             return Path.Combine(Environment.CurrentDirectory, $@"Routines\{ProjectName}\{name}.dll");
         }
 
-        private static string CompiledAssembliesPath => Path.Combine(Utilities.AssemblyDirectory, "CompiledAssemblies");
-
         private static Assembly LoadAssembly(string path)
         {
-            if (!File.Exists(path))
-            {
-                return null;
-            }
+            if (!File.Exists(path)) return null;
 
-            if (!Directory.Exists(CompiledAssembliesPath))
-            {
-                Directory.CreateDirectory(CompiledAssembliesPath);
-            }
+            if (!Directory.Exists(CompiledAssembliesPath)) Directory.CreateDirectory(CompiledAssembliesPath);
 
             var t = DateTime.Now.Ticks;
             var name = $"{Path.GetFileNameWithoutExtension(path)}{t}.{Path.GetExtension(path)}";
@@ -158,7 +208,6 @@ namespace AEAssist
             var capath = Path.Combine(CompiledAssembliesPath, name);
             Logging.Write($"Asm: {capath} origin {path}");
             if (File.Exists(capath))
-            {
                 try
                 {
                     File.Delete(capath);
@@ -167,10 +216,8 @@ namespace AEAssist
                 {
                     //
                 }
-            }
 
             if (File.Exists(pdb))
-            {
                 try
                 {
                     File.Delete(pdb);
@@ -179,17 +226,10 @@ namespace AEAssist
                 {
                     //
                 }
-            }
 
-            if (!File.Exists(capath))
-            {
-                File.Copy(path, capath);
-            }
+            if (!File.Exists(capath)) File.Copy(path, capath);
 
-            if (!File.Exists(pdb) && File.Exists(pdbPath))
-            {
-                File.Copy(pdbPath, pdb);
-            }
+            if (!File.Exists(pdb) && File.Exists(pdbPath)) File.Copy(pdbPath, pdb);
 
 
             Assembly assembly = null;
@@ -205,12 +245,12 @@ namespace AEAssist
             return assembly;
         }
 
-        void AddBehavior(Type type, string name)
+        private void AddBehavior(Type type, string name)
         {
             Behaviors.Add(name, type.GetProperty(name));
         }
 
-        void AddMethod(Type type, string name)
+        private void AddMethod(Type type, string name)
         {
             Methods.Add(name, type.GetMethod(name));
         }
@@ -232,10 +272,7 @@ namespace AEAssist
         public override void OnButtonPress()
         {
             base.OnButtonPress();
-            if (!Methods.TryGetValue("OnButtonPress", out var method))
-            {
-                return;
-            }
+            if (!Methods.TryGetValue("OnButtonPress", out var method)) return;
 
             method.Invoke(Entry, null);
         }
@@ -243,10 +280,7 @@ namespace AEAssist
         public override void Pulse()
         {
             base.Pulse();
-            if (!Methods.TryGetValue("Pulse", out var method))
-            {
-                return;
-            }
+            if (!Methods.TryGetValue("Pulse", out var method)) return;
 
             method.Invoke(Entry, null);
         }
@@ -254,104 +288,9 @@ namespace AEAssist
         public override void ShutDown()
         {
             base.ShutDown();
-            if (!Methods.TryGetValue("ShutDown", out var method))
-            {
-                return;
-            }
+            if (!Methods.TryGetValue("ShutDown", out var method)) return;
 
             method.Invoke(Entry, null);
-        }
-
-        public override Composite RestBehavior
-        {
-            get
-            {
-                if (!Behaviors.TryGetValue("RestBehavior", out var prop))
-                {
-                    return default;
-                }
-
-                return prop.GetValue(Entry, null) as Composite;
-            }
-        }
-
-
-        public override Composite PreCombatBuffBehavior
-        {
-            get
-            {
-                if (!Behaviors.TryGetValue("PreCombatBuffBehavior", out var prop))
-                {
-                    return default;
-                }
-
-                return prop.GetValue(Entry, null) as Composite;
-            }
-        }
-
-        public override Composite PullBehavior
-        {
-            get
-            {
-                if (!Behaviors.TryGetValue("PullBehavior", out var prop))
-                {
-                    return default;
-                }
-
-                return prop.GetValue(Entry, null) as Composite;
-            }
-        }
-
-        public override Composite HealBehavior
-        {
-            get
-            {
-                if (!Behaviors.TryGetValue("HealBehavior", out var prop))
-                {
-                    return default;
-                }
-
-                return prop.GetValue(Entry, null) as Composite;
-            }
-        }
-
-        public override Composite CombatBuffBehavior
-        {
-            get
-            {
-                if (!Behaviors.TryGetValue("CombatBuffBehavior", out var prop))
-                {
-                    return default;
-                }
-
-                return prop.GetValue(Entry, null) as Composite;
-            }
-        }
-
-        public override Composite CombatBehavior
-        {
-            get
-            {
-                if (!Behaviors.TryGetValue("CombatBehavior", out var prop))
-                {
-                    return default;
-                }
-
-                return prop.GetValue(Entry, null) as Composite;
-            }
-        }
-
-        public override Composite PullBuffBehavior
-        {
-            get
-            {
-                if (!Behaviors.TryGetValue("PullBuffBehavior", out var prop))
-                {
-                    return default;
-                }
-
-                return prop.GetValue(Entry, null) as Composite;
-            }
         }
     }
 }
